+++ airgeddon.sh
@@ -2,7 +2,7 @@
 #Title........: airgeddon.sh
 #Description..: This is a multi-use bash script for Linux systems to audit wireless networks.
 #Author.......: v1s1t0r
-#Date.........: 20180926
+#Date.........: 20181013
 #Version......: 8.12
 #Usage........: bash airgeddon.sh
 #Bash Version.: 4.2 or later
@@ -10619,15 +10619,17 @@ function initialize_script_settings() {
 	declare -gA wps_data_array
 }
 
-#Detect if there is a working X window system excepting for docker container
+#Detect if there is a working X window system excepting for docker container and wayland
 function check_xwindow_system() {
 
 	debug_print
 
 	if hash xset 2> /dev/null; then
 		if ! xset -q > /dev/null 2>&1; then
-			if [ "${is_docker}" -eq 0 ]; then
-				xterm_ok=0
+			if [ "${XDG_SESSION_TYPE}" != "wayland" ]; then
+				if [ "${is_docker}" -eq 0 ]; then
+					xterm_ok=0
+				fi
 			fi
 		fi
 	fi