+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Improved regex for BSSID check
  - Changes on contributing policies
  - Github pull request template added and issue templates updated
+ - Added Wayland graphic system support (Thank you to "fr0xk")
 
 ### 8.11
  - Turkish language translation added (Thank you to "UN5T48L3")